+++ src/createComponent.ts
@@ -9,23 +9,36 @@ import statefulWidget from './templates/statefulWidget';
 import clazz from './templates/clazz';
 import interfaceClazz from './templates/interfaceClazz';
 import mobxStore from './templates/mobxStore';
-import { camelCase, kebabCase, snakeCase, upperFirst } from 'lodash';
+import { camelCase, kebabCase, lowerCase, snakeCase, upperFirst } from 'lodash';
 import pascalCase from './templates/shared/functions/pascal-case';
 import clazzImplementation from './templates/clazzImplementation';
 import getxFeatureBinding from './templates/getxFeatureBinding';
 import getxFeatureController from './templates/getxFeatureController';
 import getxFeatureView from './templates/getxFeatureView';
 import getxFeatureRoutes from './templates/getxFeatureRoutes';
+import getxAppBindings from './templates/getxAppBindings';
+import getxAppPages from './templates/getxAppPages';
+import getxService from './templates/getxService';
+import flutterMain from './templates/flutterMain';
 
 interface ComponentProps {
   dir?: string;
-  type: 'widget' | 'class' | 'model' | 'controller' | 'interface' | 'provider' | 'repository' | 'service' | 'getx feature' | 'getx route' | 'store';
+  type: 'widget' | 'class' | 'dto'  | 'model' | 'controller' | 'interface' | 'provider' | 'repository' | 'service' | 'getx-feature' | 'getx-route' |'getx-service' | 'getx-structure' | 'mobx-store';
   stateFullWidget?: boolean;
+}
+
+interface GetxFeature {
+  dir: string;
+  fileName: string;
+  componentName: string;
+  getxViewsSuffix: string;
+  fromRoot?: boolean;
 }
 
 export default async (componentName: string, { dir, type, stateFullWidget = false }: ComponentProps) => {
   // Load configurations.
   const config = vscode.workspace.getConfiguration("createFlutterWidgetsAndClasses");
+  const getxViewsSuffix = config.get("getxViewsSuffix") as string;
   const useIPrefixForInterfaces = config.get("useIPrefixForInterfaces") as boolean;
   const createImplementationOfInterface = config.get("createImplementationOfInterface") as boolean;
 
@@ -51,19 +64,24 @@ export default async (componentName: string, { dir, type, stateFullWidget = fals
   } else if (type === 'controller') {
     componentName += 'Controller';
     componentFileName = `${ fileName }_controller.dart`;
+  } else if (type === 'dto') {
+    componentName += 'DTO';
+    componentFileName = `${ fileName }_dto.dart`;
   } else if (type === 'model') {
     componentName += 'Model';
     componentFileName = `${ fileName }_model.dart`;
-  } else if (type === 'getx route') {
+  } else if (type === 'getx-route') {
     componentFileName = `${ fileName }_routes.dart`;
   } else {
     componentFileName = `${ fileName }.dart`;
   }
 
+  const sep = path.sep;
+
   if (!dir) {
     dir =
       (await vscode.window.showInputBox({
-        value: "/",
+        value: sep,
         prompt: `Path from root`,
         ignoreFocusOut: true,
         valueSelection: [-1, -1]
@@ -74,24 +92,38 @@ export default async (componentName: string, { dir, type, stateFullWidget = fals
     dir = projectRoot + dir;
   }
 
-  if (dir[dir.length - 1] !== "/") {
-    dir = dir + "/";
+  if (dir[dir.length - 1] !== sep) {
+    dir = dir + sep;
   }
 
-  let dirWithFileName: string;
-
-
   const filePath = (fileName: string) => {
-    return dir + "/" + fileName;
-
-    // return dirWithFileName + "/" + fileName;
+    return dir + sep + fileName;
   }
 
   const filePathFeature = (folderName: string, fileName: string) => {
-    return dir + "/" + folderName + "/" + fileName;
-
-    // return dirWithFileName + "/" + fileName;
+    // return dir + sep + folderName + sep + fileName;
+    return folderName + sep + fileName;
   }
+
+  const createGetxFeature = async ({
+    dir, fileName, componentName, getxViewsSuffix, fromRoot = false,
+  }: GetxFeature) => {
+    let pathDir: string = `${dir}${sep}${fileName}`;
+
+    await mkdirp(pathDir);
+
+    await createFile(
+      filePathFeature(pathDir, `${ fileName }_binding.dart`), getxFeatureBinding({ componentName, fileName })
+    );
+
+    await createFile(
+      filePathFeature(pathDir, `${ fileName }_controller.dart`), getxFeatureController({ componentName, fileName })
+    );
+
+    await createFile(
+      filePathFeature(pathDir, `${ fileName }_${lowerCase(getxViewsSuffix)}.dart`), getxFeatureView({ componentName, fileName, getxViewsSuffix })
+    );
+  };
 
   if (type === 'widget' && !stateFullWidget) {
     await createFile(
@@ -105,15 +137,21 @@ export default async (componentName: string, { dir, type, stateFullWidget = fals
     );
   }
 
-  if (type === 'class' || type === 'controller' || type === 'model') {
+  if (type === 'class' || type === 'controller' || type === 'model' || type === 'dto') {
     await createFile(
       filePath(componentFileName), clazz({ componentName })
     );
   }
 
-  if (type === 'getx route') {
+  if (type === 'getx-route') {
     await createFile(
-      filePath(componentFileName), getxFeatureRoutes({ componentName })
+      filePath(componentFileName), getxFeatureRoutes({ componentName, getxViewsSuffix })
+    );
+  }
+
+  if (type === 'getx-service') {
+    await createFile(
+      filePath(componentFileName), getxService({ componentName })
     );
   }
 
@@ -131,31 +169,76 @@ export default async (componentName: string, { dir, type, stateFullWidget = fals
     }
   }
 
-  if (type === 'getx feature' ) {
-    await mkdirp(dir + '/' + fileName);
+  if (type === 'getx-feature') {
+    await createGetxFeature({ dir, fileName, componentName, getxViewsSuffix });
+  }
 
+  if (type === 'getx-structure') {
+    const mainFile = dir + sep + 'main.dart';
+    const oldMainFile = dir + sep + 'remove_old_main.dart';
+
+    fs.rename(mainFile, oldMainFile, () => null);
+
+    await mkdirp(dir + '/app');
+    await mkdirp(dir + '/app/core');
+
+    await mkdirp(dir + '/app/core/bindings');
     await createFile(
-      filePathFeature(fileName, `${ fileName }_binding.dart`), getxFeatureBinding({ componentName, fileName })
+      filePath(`app/core/bindings/application_bindings.dart`), getxAppBindings({ componentName: 'application_bindings' })
+    );
+
+    await mkdirp(dir + '/app/core/config');
+    await mkdirp(dir + '/app/core/theme');
+    await mkdirp(dir + '/app/core/widgets');
+    await mkdirp(dir + '/app/data');
+    await mkdirp(dir + '/app/data/dtos');
+    await mkdirp(dir + '/app/data/enums');
+    await mkdirp(dir + '/app/data/exceptions');
+    await mkdirp(dir + '/app/data/extensions');
+    await mkdirp(dir + '/app/data/models');
+    await mkdirp(dir + '/app/data/providers');
+    await mkdirp(dir + '/app/data/repositories');
+    await mkdirp(dir + '/app/data/services');
+    await mkdirp(dir + '/app/data/sessions');
+
+    await mkdirp(dir + '/app/modules');
+    await createGetxFeature({
+      dir: `${dir}app${sep}modules`,
+      fileName: 'home',
+      componentName: 'home',
+      getxViewsSuffix,
+      fromRoot: true,
+    });
+
+    await mkdirp(dir + '/app/routes');
+    await createFile(
+      filePath('app/routes/app_pages.dart'), getxAppPages({ componentName: 'app_pages' })
+    );
+    await createFile(
+      filePath('app/routes/home_routes.dart'), getxFeatureRoutes({ componentName: 'home', getxViewsSuffix, createHomeImport: true })
     );
 
     await createFile(
-      filePathFeature(fileName, `${ fileName }_controller.dart`), getxFeatureController({ componentName, fileName })
-    );
-
-    await createFile(
-      filePathFeature(fileName, `${ fileName }_view.dart`), getxFeatureView({ componentName, fileName })
+      filePath('main.dart'), flutterMain({ componentName })
     );
   }
 
-  if (type === 'store' ) {
+  if (type === 'mobx-store' ) {
     await createFile(
       filePath(componentFileName), mobxStore({ componentName, fileName })
     );
   }
 
   setTimeout(() => {
-    if (type === 'getx feature') {
-      vscode.workspace.openTextDocument(filePath(`${fileName}/${ fileName }_view.dart`)).then(editor => {
+    if (type === 'getx-feature') {
+      vscode.workspace.openTextDocument(filePath(`${fileName}/${ fileName }_${lowerCase(getxViewsSuffix)}.dart`)).then(editor => {
+        if (!editor) {
+          return;
+        }
+        vscode.window.showTextDocument(editor);
+      });
+    } else if ((type === 'getx-structure')) {
+      vscode.workspace.openTextDocument(filePath(`app/modules/home/home_${lowerCase(getxViewsSuffix)}.dart`)).then(editor => {
         if (!editor) {
           return;
         }