+++ src/templates/flutterMain.ts
@@ -1,0 +1,24 @@
+import { kebabCase, snakeCase } from 'lodash';
+import CreateComponent from './interfaces/CreateComponent';
+import pascalCase from './shared/functions/pascal-case';
+
+export default ({ componentName }: CreateComponent) => (
+`import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+
+import 'app/core/bindings/application_bindings.dart';
+import 'app/routes/app_pages.dart';
+
+void main() {
+  runApp(
+    GetMaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Your App Title',
+      initialBinding: ApplicationBindings(),
+      initialRoute: AppPages.INITIAL,
+      getPages: AppPages.routes,
+    ),
+  );
+}
+`
+);