+++ src/templates/getxService.ts
@@ -1,0 +1,11 @@
+import CreateComponent from './interfaces/CreateComponent';
+import pascalCase from './shared/functions/pascal-case';
+
+export default ({ componentName }: CreateComponent) => (
+`import 'package:get/get.dart';
+
+class ${ pascalCase(componentName) } extends GetxService {
+  //TODO: Implement ${ pascalCase(componentName) } GetxService.
+}
+`
+);