+++ src/templates/interfaces/CreateComponent.ts
@@ -1,5 +1,7 @@
 export default interface CreateComponent {
   componentName: string;
   fileName?: string;
+  getxViewsSuffix?: string;
   useIPrefixForInterfaces?: boolean;
+  createHomeImport?: boolean;
 }